feat: extracts GridStylesheet into its own component
