+++ src/components/GridStylesheet.jsx
@@ -1,0 +1,25 @@
+import { useContext, useEffect, useRef } from "preact/hooks";
+import GridSystemContext from "./GridSystemContext";
+
+/**
+ * Injects generated grid CSS into the document head.
+ *
+ * Creates a style element on mount and updates its content whenever
+ * the grid CSS changes in the context.
+ *
+ * @returns {null} This component doesn't render any visible output.
+ */
+export default function GridStylesheet() {
+  const { gridCss } = useContext(GridSystemContext);
+  const styleRef = useRef(null);
+
+  useEffect(() => {
+    if (!styleRef.current) {
+      const style = document.createElement("style");
+      document.head.appendChild(style);
+      styleRef.current = style;
+    }
+
+    styleRef.current.textContent = gridCss;
+  }, [gridCss]);
+}
